First iteration of refactoring. Moved models into types, deleted GraphQL directory. Next I'm gonna convert controllers into resolvers. Then I'll see how I want to hook up the REST version.
