+++ src/main.rs
@@ -14,8 +14,8 @@ extern crate r2d2;
 extern crate rand;
 
 use crate::db::Connection;
-use crate::graphql::Context;
-use crate::types::Result;
+use crate::types::Context;
+use crate::utils::Result;
 use rocket::response::content;
 use rocket::*;
 use rocket_cors::CorsOptions;
@@ -23,12 +23,11 @@ use rocket_cors::CorsOptions;
 mod controllers;
 mod db;
 mod github;
-mod graphql;
-mod models;
 mod routes;
 mod schema;
 mod tokens;
 mod types;
+mod utils;
 
 #[get("/")]
 fn index() -> String {
@@ -45,7 +44,7 @@ fn handle_graphql_get(
     request: juniper_rocket::GraphQLRequest,
     database: Connection,
 ) -> juniper_rocket::GraphQLResponse {
-    let schema = graphql::create_schema();
+    let schema = types::create_schema();
     let context = Context { database };
     request.execute(&schema, &context)
 }
@@ -55,7 +54,7 @@ fn handle_graphql_post(
     request: juniper_rocket::GraphQLRequest,
     database: Connection,
 ) -> juniper_rocket::GraphQLResponse {
-    let schema = graphql::create_schema();
+    let schema = types::create_schema();
     let context = Context { database };
     request.execute(&schema, &context)
 }
