--- src/models/mod.rs
@@ -1,6 +1,0 @@
-mod project;
-mod submission;
-mod user;
-pub use self::project::*;
-pub use self::submission::*;
-pub use self::user::*;
