+++ src/types/user.rs
@@ -1,10 +1,19 @@
 use crate::schema::*;
-use crate::types::Result;
+use crate::utils::Result;
+use super::Context;
+use crate::types::{Submission, Project};
+use crate::schema::{projects, submissions};
 use argonautica::input::Salt;
 use argonautica::{Hasher, Verifier};
 use diesel::{self, AsChangeset, Queryable};
 use serde_derive::{Deserialize, Serialize};
 use std::env;
+use diesel::pg::expression::dsl::any;
+use diesel::BelongingToDsl;
+use diesel::ExpressionMethods;
+use diesel::QueryDsl;
+use diesel::RunQueryDsl;
+
 
 const SALT_LENGTH: u32 = 16;
 
@@ -70,15 +79,34 @@ pub struct UserResponse {
     pub email: String,
 }
 
+graphql_object!(UserResponse: Context as "User" |&self| {
+    description: "A user"
+
+    field id(&executor) -> i32 {
+        self.id
+    }
+
+    field display_name(&executor) -> &str {
+        &self.display_name
+    }
+
+    field email(&executor) -> &str {
+        &self.email
+    }
+    
+    field projects(&executor) -> Vec<Project> {
+        let database: &diesel::PgConnection = &executor.context().database;
+        let project_ids = Submission::belonging_to(self).select(submissions::project_id);
+        projects::table
+            .filter(projects::id.eq(any(project_ids)))
+            .load::<Project>(database).expect("Could not load projects")
+    }
+});
+
 #[derive(Serialize, Deserialize, GraphQLObject)]
 pub struct LoginRequest {
     pub email: String,
     pub password: String,
-}
-
-pub struct LoginResponse {
-    pub token: String,
-    pub user: UserResponse,
 }
 
 impl UserResponse {
